requirements.txt
torch==2.6.0+cpu
torchvision==0.21.0+cpu

pytorch-lightning==2.6.0

lightly==1.5.22

numpy==2.2.6
pandas==2.2.2
scikit-learn==1.3.2
Pillow==10.4.0

matplotlib==3.9.1